% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_details.R
\name{cqc_location_details}
\alias{cqc_location_details}
\title{Location Details}
\usage{
cqc_location_details(x, location_col = "location_id", verbose = TRUE,
  clean_names = TRUE)
}
\arguments{
\item{x}{A data frame with location IDs, e.g. one
returned by \code{\link[=cqc_location_search]{cqc_location_search()}}.}

\item{location_col}{The name of the column with location IDs, defaults to
\code{"location_id"}.}

\item{verbose}{If \code{TRUE} prints download progress to console if requesting
more than 1,000 records. Defaults to \code{TRUE}.}

\item{clean_names}{If \code{TRUE}, converts a column names to snake_case.
Defaults to \code{TRUE}.}
}
\value{
A list with location details
}
\description{
Location Details
}
