% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partner_code.R
\name{cqc_partner_code}
\alias{cqc_partner_code}
\title{Partner Code}
\usage{
cqc_partner_code(check_env = FALSE)
}
\arguments{
\item{check_env}{If TRUE, will check the environment variable
\code{CQC_PARTNER_CODE} first before asking for user input.}
}
\description{
Function to add an identifier code to API queries.
}
\details{
The CQC wants all organisations using this API to add an additional query
identifying themselves to all queries.

If you are a CQC Syndication partner the value for this code
will be provided to you, otherwise an informative but concise code
representing your organisation should be chosen. The partner code can be set
as an environmental variable named \code{CQC_PARTNER_CODE}, or using this
function. If set, it will automatically be included in all queries.
}
