% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{cqc_reports}
\alias{cqc_reports}
\title{Title}
\usage{
cqc_reports(inspection_report_link_id, related_document_type = NULL,
  plain_text = TRUE, verbose = TRUE)
}
\arguments{
\item{verbose}{}
}
\description{
Title
}
\examples{

}
