% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscription_key.R
\name{cqc_subscription_key}
\alias{cqc_subscription_key}
\title{API Subscription Key}
\usage{
cqc_subscription_key(check_env = FALSE)
}
\arguments{
\item{check_env}{If TRUE, will check the environment variable
\code{CQC_SUBSCRIPTION_KEY} first before asking for user input.}
}
\description{
Function to add a subscription key to API queries.
}
\details{
The CQC requires all organisations using this API to obtain
an API key from their website and append to queries

According to the CQC's documentation,
Users will now need to authenticate.
Authentication helps us understand how people use our API
and make sure that it fits with your needs.
It also means that we can tell you about any future changes.
}
