% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqcr-package.R
\name{cqcr}
\alias{cqcr-package}
\title{cqcr}
\description{
Access data from the 'Care Quality Commission', the health and adult social
care regulator for England. The 'Care Quality Commission' operates an
\href{https://www.cqc.org.uk/about-us/transparency/using-cqc-data#api}{API},
with data available under the Open Government License. Data includes
information on service providers, locations such as hospitals, care homes
and medical clinics, and ratings and inspection reports. Registration is
not required, although you should use the \code{\link[=cqc_partner_code]{cqc_partner_code()}} function
to set an indicator of the organisation you are from when querying the API.
}
