% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspection_area.R
\name{cqc_inspection_areas}
\alias{cqc_inspection_areas}
\title{Inspection areas}
\usage{
cqc_inspection_areas(clean_names = TRUE)
}
\arguments{
\item{clean_names}{If \code{TRUE}, converts a column names to snake_case.
Defaults to \code{TRUE}.}
}
\value{
A tibble with inspection areas.
}
\description{
Retrieves all CQC inspection areas.
}
