% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/providers_basic.R
\name{cqc_provider}
\alias{cqc_provider}
\alias{cqc_provider_locations}
\alias{cqc_provider_inspection_areas}
\title{Individual providers}
\usage{
cqc_provider(provider_id, clean_names = TRUE)

cqc_provider_locations(provider_id, clean_names = TRUE)

cqc_provider_inspection_areas(provider_id, clean_names = TRUE)
}
\arguments{
\item{clean_names}{If \code{TRUE}, converts a column names to snake_case.
Defaults to \code{TRUE}.}
}
\description{
Return information on individual providers
}
